+++ apis/openweathermap.js
@@ -1,42 +1,38 @@
 /*jshint esversion: 6,node: true,-W041: false */
 "use strict";
 
-const Openweathermap = require('openweather-apis'),
-	converter = require('../util/converter'),
+const converter = require('../util/converter'),
 	moment = require('moment-timezone'),
 	geoTz = require('geo-tz'),
+	request = require('request'),
 	debug = require('debug')('homebridge-weather-plus');
 
 class OpenWeatherMapAPI
 {
-	constructor(apiKey, language, location, locationGeo, locationCity, conditionDetail, log)
+	constructor(apiKey, language, locationId, locationGeo, locationCity, conditionDetail, log)
 	{
-		Openweathermap.setLang(language);
-		if (location)
-		{
-			Openweathermap.setCityId(location);
-		}
-		if (locationGeo)
-		{
-			Openweathermap.setCoordinate(locationGeo[0], locationGeo[1]);
-		}
-		if (locationCity)
-		{
-			Openweathermap.setCity(locationCity);
-		}
-		// HomeKit expects Celsius
-		Openweathermap.setUnits('metric');
-		Openweathermap.setAPPID(apiKey);
+		this.log = log;
+		this.apiKey = apiKey;
+
+		this.locationId = locationId;
+		this.locationCity = locationCity;
+		this.locationGeo = locationGeo;
+
 		this.attribution = 'Powered by OpenWeatherMap';
 		this.reportCharacteristics = [
 			'AirPressure',
+			'CloudCover',
 			'Condition',
 			'ConditionCategory',
+			'DewPoint',
 			'Humidity',
 			'ObservationTime',
+			'Rain1h',
 			'RainBool',
 			'SnowBool',
 			'Temperature',
+			'TemperatureWindChill',
+			'UVIndex',
 			'WindDirection',
 			'WindSpeed'
 		];
@@ -45,90 +41,86 @@ class OpenWeatherMapAPI
 			'CloudCover',
 			'Condition',
 			'ConditionCategory',
+			'DewPoint',
 			'ForecastDay',
 			'Humidity',
 			'RainBool',
+			'RainDay',
 			'SnowBool',
 			'TemperatureMax',
 			'TemperatureMin',
+			'TemperatureWindChill',
+			'UVIndex',
 			'WindDirection',
 			'WindSpeed'
 		];
-		this.forecastDays = 5;
+		this.forecastDays = 8;
 		this.conditionDetail = conditionDetail;
-		this.log = log;
 	}
 
 	update(forecastDays, callback)
 	{
 		debug("Updating weather with OpenWeatherMap");
+		let that = this;
 
-		let weather = {};
-		Openweathermap.getAllWeather(function (err, jsonObj)
+		if (!this.locationGeo)
 		{
-			if (!err && jsonObj.coord)
+			this.getLocationGeo((error, coordinates) =>
 			{
-				// Current weather report
-				this.parseReport(weather, jsonObj, callback);
-			}
-			else
-			{
-				this.log.error("Error retrieving weather report");
-				this.log.error("Error Message: " + (err ? err : JSON.stringify(jsonObj)));
-				callback(err ? err : true);
-			}
-		}.bind(this));
+				if (!error)
+				{
+					that.locationGeo = [coordinates.lat, coordinates.lon];
+					debug("Found locationGeo: %s", that.locationGeo);
+					that.update(forecastDays, callback);
+				}
+				else
+				{
+					that.log.error("Error getting locationGeo from %s", this.locationId ? this.locationId : this.locationCity);
+					that.log.error(error);
+				}
+			});
+		}
+		else
+		{
+			debug("Update weather");
+			let weather = {};
 
-		Openweathermap.getWeatherForecast(function (err, jsonObj)
-		{
-			if (!err && jsonObj.city)
+			this.getWeatherData((error, result) =>
 			{
-				this.parseForecasts(weather, jsonObj, callback);
-			}
-			else
-			{
-				this.log.error("Error retrieving weather forecast for OpenWeatherMap");
-				this.log.error("Error Message: " + (err ? err : JSON.stringify(jsonObj)));
-				callback(err ? err : true);
-			}
-		}.bind(this));
+				if (!error)
+				{
+					this.parseReport(weather, result["current"], result["timezone"], callback);
+					this.parseForecasts(weather, result["daily"], result["timezone"], callback);
+				}
+				else
+				{
+					that.log.error("Error retrieving weather report");
+					that.log.error(error);
+					callback();
+				}
+			});
+		}
 	}
 
-	parseReport(weather, values, callback)
+	parseReport(weather, values, timezone, callback)
 	{
-		let report = weather.report || {};
-		const timezone = String(geoTz(values.coord.lat, values.coord.lon));
+		weather.report = {};
+		this.parseWeather(weather.report, values);
+		weather.report.ObservationTime = moment.unix(values.dt).tz(timezone).format('HH:mm:ss');
+		weather.report.Rain1h = isNaN(parseInt(values.rain)) ? 0 : parseInt(values.rain);
 
-		report.AirPressure = parseInt(values.main.pressure);
-		report.CloudCover = parseInt(values.clouds.all);
-		report.Condition = values.weather[0].description;
-		report.ConditionCategory = this.getConditionCategory(values.weather[0].id, this.conditionDetail);
-		report.Humidity = parseInt(values.main.humidity);
-		report.ObservationTime = moment.unix(values.dt).tz(timezone).format('HH:mm:ss');
-		let detailedCondition = this.getConditionCategory(values.weather[0].id, true);
-		report.RainBool = [5,6].includes(detailedCondition);
-		report.SnowBool = [7,8].includes(detailedCondition);
-		report.Temperature = values.main.temp;
-		report.WindDirection = converter.getWindDirection(values.wind.deg);
-		report.WindSpeed = values.wind.speed;
-
-		weather.report = report;
 		if (weather.forecasts)
 		{
 			callback(null, weather);
 		}
 	}
 
-	parseForecasts(weather, values, callback)
+	parseForecasts(weather, values, timezone, callback)
 	{
-		const timezone = String(geoTz(values.city.coord.lat, values.city.coord.lon));
-
 		let forecasts = [];
-		// We get a forecast for 5 days with values each 3 hours.
-		const itemsFiltered = this.prepareForecasts(values.list, timezone);
-		for (let i = 0; i < itemsFiltered.length; i++)
+		for (let i = 0; i < values.length; i++)
 		{
-			forecasts[forecasts.length] = this.parseForecast(itemsFiltered[i], timezone);
+			forecasts[forecasts.length] = this.parseForecast(values[i], timezone);
 		}
 
 		weather.forecasts = forecasts;
@@ -141,73 +133,33 @@ class OpenWeatherMapAPI
 	parseForecast(values, timezone)
 	{
 		let forecast = {};
-
-		// TODO implement rainbool and snowbool
-		forecast.AirPressure = parseInt(values.main.pressure);
-		forecast.CloudCover = parseInt(values.clouds.all);
-		forecast.Condition = values.weather[0].description;
-		forecast.ConditionCategory = this.getConditionCategory(values.weather[0].id, this.conditionDetail);
+		this.parseWeather(forecast, values);
 		forecast.ForecastDay = moment.unix(values.dt).tz(timezone).format('dddd');
-		forecast.Humidity = parseInt(values.main.humidity);
-		let detailedCondition = this.getConditionCategory(values.weather[0].id, true);
-		forecast.RainBool = [5,6].includes(detailedCondition);
-		forecast.SnowBool = [7,8].includes(detailedCondition);
-		forecast.TemperatureMax = values.main.temp_max;
-		forecast.TemperatureMin = values.main.temp_min;
-		forecast.WindDirection = converter.getWindDirection(values.wind.deg);
-		forecast.WindSpeed = values.wind.speed;
+		forecast.RainDay = isNaN(parseInt(values.rain)) ? 0 : parseInt(values.rain);
+		forecast.TemperatureMax = values.temp.max;
+		forecast.TemperatureMin = values.temp.min;
 
 		return forecast;
 	}
 
-	/**
-	 * OpenWeatherMap gives free users a 5-day forecast BUT with one value each 3 hours....
-	 * This method searches the value at around noon so only one value for one day is returned.
-	 * The first value in the returned array is the forecase for today.
-	 * Actually this method should somehow e.g. average humidity, find mix/max, ... for all values for
-	 * one day to give an accurate result.
-	 *
-	 * @param forecasts
-	 * @param timezone
-	 * @return {Array}
-	 */
-	prepareForecasts(forecasts, timezone)
+	parseWeather(report, values)
 	{
-		let forecastsFiltered = [];
-		const endOfToday = moment.tz(timezone).endOf("day").unix();
-		// Find one for today
-		for (let i = 0; i < forecasts.length; i++)
-		{
-			const listItem = forecasts[i];
-			if (listItem.dt <= endOfToday)
-			{
-				const hourOfDay = moment.unix(listItem.dt).tz(timezone).format("H");
-				if (parseInt(hourOfDay) >= 10 && parseInt(hourOfDay) <= 13)
-				{
-					forecastsFiltered[forecastsFiltered.length] = forecasts[i];
-				}
-			}
-		}
-		// The request was made after noon, take first for today
-		if (forecastsFiltered.length === 0)
-		{
-			forecastsFiltered[forecastsFiltered.length] = forecasts[0];
-		}
-
-		// For all following days find exactly one value (best: noon)
-		for (let i = 0; i < forecasts.length; i++)
-		{
-			const listItem = forecasts[i];
-			if (listItem.dt > endOfToday)
-			{
-				const hourOfDay = moment.unix(listItem.dt).tz(timezone).format("H");
-				if (parseInt(hourOfDay) >= 10 && parseInt(hourOfDay) <= 13)
-				{
-					forecastsFiltered[forecastsFiltered.length] = forecasts[i];
-				}
-			}
-		}
-		return forecastsFiltered;
+		report.AirPressure = parseInt(values.pressure);
+		report.CloudCover = parseInt(values.clouds);
+		report.Condition = values.weather[0].description;
+		report.ConditionCategory = this.getConditionCategory(values.weather[0].id, this.conditionDetail);
+		report.DewPoint = parseInt(values.dew_point);
+		report.Humidity = parseInt(values.humidity);
+		let detailedCondition = this.getConditionCategory(values.weather[0].id, true);
+		report.RainBool = [5,6].includes(detailedCondition);
+		report.SnowBool = [7,8].includes(detailedCondition);
+		debug(values.temp);
+		debug(typeof values.temp === 'object');
+		report.Temperature = typeof values.temp === 'object' ? parseInt(values.temp.day) : parseInt(values.temp);
+		report.TemperatureWindChill = typeof values.feels_like === 'object' ? parseInt(values.feels_like.day) : parseInt(values.feels_like);
+		report.UVIndex = parseInt(values.uvi);
+		report.WindDirection = converter.getWindDirection(values.wind_deg);
+		report.WindSpeed = values.wind_speed;
 	}
 
 	getConditionCategory(code, detail = false)
@@ -268,6 +220,69 @@ class OpenWeatherMapAPI
 			this.log.warn("Unknown OpenWeatherMap category " + code);
 			return 0;
 		}
+	};
+
+	getWeatherData(callback)
+	{
+		debug("Getting weather data for location %s", this.locationGeo);
+
+		const queryUri = "https://api.openweathermap.org/data/2.5/onecall?units=metric&lat=" + this.locationGeo[0] + "&lon=" + this.locationGeo[1] + "&appid=" + this.apiKey;
+		request(encodeURI(queryUri),  (requestError, response, body) =>
+		{
+			if (!requestError)
+			{
+				let parseError;
+				let weather
+				try
+				{
+					weather = JSON.parse(body);
+				} catch (e)
+				{
+					parseError = e;
+				}
+				callback(parseError, weather);
+			}
+			else
+			{
+				callback(requestError);
+			}
+		});
+	}
+
+	getLocationGeo(callback)
+	{
+		debug("Getting locationGeo from %s", this.locationId ? this.locationId : this.locationCity);
+
+		let locationQuery = this.locationId
+			? "id=" + this.locationId
+			: "q=" + this.locationCity.toLowerCase();
+
+		const queryUri = "https://api.openweathermap.org/data/2.5/weather?" + locationQuery + "&appid=" + this.apiKey;
+		request(encodeURI(queryUri),  (requestError, response, body) =>
+		{
+			if (!requestError)
+			{
+				// Get locationGeo from weather report
+				let coordinates;
+				let parseError;
+				try
+				{
+					coordinates = JSON.parse(body)["coord"];
+					if (coordinates === undefined)
+					{
+						parseError = body;
+					}
+				} catch (e)
+				{
+					parseError = e;
+				}
+				callback(parseError, coordinates);
+			}
+			else
+			{
+				callback(requestError);
+			}
+		});
 	};
 }