+++ util/characteristics.js
@@ -25,6 +25,7 @@ const inherits = require('util').inherits,
 		SnowBool: 'f14eb1ad-e000-4ce6-bd0e-384f9ec4d5dd',
 		SolarRadiation: '1819a23e-ecab-4d39-b29a-7364d299310b',
 		TemperatureMin: '707b78ca-51ab-4dc9-8630-80a58f07e419',
+		TemperatureWindChill: 'c1283352-3d12-4777-acd5-4734760f1ac8',
 		UVIndex: '05ba0fe0-b848-4226-906d-5b64272e05ce',
 		Visibility: 'd24ecc1e-6fad-4fb5-8137-5af88bd5e857',
 		WindDirection: '46f1284c-1912-421b-82f5-eb75008b167e',
@@ -351,6 +352,14 @@ module.exports = function (Characteristic, units)
 	};
 	inherits(CustomCharacteristic.TemperatureMin, Characteristic);
 	// Homekit converts temperature by itself accoding to the user device settings
+
+	CustomCharacteristic.TemperatureWindChill = function ()
+	{
+		Characteristic.call(this, 'Temperature Wind Chill', CustomUUID.TemperatureWindChill);
+		this.setProps(temperatureProps(50, -50));
+		this.value = this.getDefaultValue();
+	};
+	inherits(CustomCharacteristic.TemperatureWindChill, Characteristic);
 
 	CustomCharacteristic.UVIndex = function ()
 	{